Renamed component files to start with caps, changed gameboard to be its own component
